+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Ansible-Alicloud Change Log
 
 ## 1.7.0 (Unreleased)
+
+IMPROVEMENTS:
+
+- Improve ali_security_group and support purge_rules and purge_rules_egress ([#123](https://github.com/alibaba/ansible-provider/pull/123))
+
 ## 1.6.0 (October 30, 2018)
 
 IMPROVEMENTS: